requirements.txt
Flask
flask-cors
gunicorn
pyyaml
